feat: Add login functionality and retrieve additional user information
This commit adds the login functionality to the index.php file. It checks if the password is correct and returns the user's information if it is. Additionally, it retrieves the user's additional information from the user_additional_info table and returns it along with the user's basic information. If the user does not have additional information, an error message is returned.

The commit also includes a new function, getUserAdditionalInfo(), which retrieves the additional information of a user based on their email.

This commit addresses the need to implement login functionality and retrieve additional user information.